+++ blueprints/lib/models/state.js
@@ -1,0 +1,8 @@
+module.exports = {
+  statesEnum: {
+    NEW: 'new',
+    NEED_UPDATE: 'need update',
+    INSTALLED: 'installed',
+    MANDATORY: 'mandatory'
+  }
+}
