--- blueprints/models/actions-enum.js
@@ -1,6 +1,0 @@
-module.exports = {
-  OVERWRITE: 'overwrite',
-  SKIP: 'skip',
-  IDENTICAL: 'identical',
-  DIFF: 'diff'
-}
