--- blueprints/models/states-enum.js
@@ -1,6 +1,0 @@
-module.exports = {
-  NEW: 'new',
-  NEED_UPDATE: 'need update',
-  INSTALLED: 'installed',
-  MANDATORY: 'mandatory'
-}
